refactor: 内部模块加 _ 前缀，仅通过 __init__.py 暴露公开 API
