+++ packages/diydev/src/diydev/__init__.py
@@ -1,0 +1,5 @@
+"""diydev — diy 开发流程 CLI。"""
+
+from ._cli import main
+
+__all__ = ["main"]
