+++ packages/diydev/src/diydev/_cli.py
@@ -4,14 +4,14 @@ import sys
 from typing import Annotated
 from cyclopts import App, Parameter, Group
 
-from .log import VerboseFlag, set_verbosity
-from .list_cmd import work_list
-from .start_cmd import start_app
-from .new_cmd import new_app
-from .status_cmd import status_app
-from .pr_cmd import pr_app
-from .merge_pr_cmd import merge_pr_app
-from .prune_cmd import prune_app
+from ._log import VerboseFlag, set_verbosity
+from ._list_cmd import work_list
+from ._start_cmd import start_app
+from ._new_cmd import new_app
+from ._status_cmd import status_app
+from ._pr_cmd import pr_app
+from ._merge_pr_cmd import merge_pr_app
+from ._prune_cmd import prune_app
 
 _GLOBAL_GROUP = Group("全局选项")
