+++ packages/diydev/src/diydev/_work.py
@@ -2,13 +2,13 @@
 
 from cyclopts import App
 
-from .list_cmd import work_list
-from .start_cmd import start_app
-from .new_cmd import new_app
-from .status_cmd import status_app
-from .pr_cmd import pr_app
-from .merge_pr_cmd import merge_pr_app
-from .prune_cmd import prune_app
+from ._list_cmd import work_list
+from ._start_cmd import start_app
+from ._new_cmd import new_app
+from ._status_cmd import status_app
+from ._pr_cmd import pr_app
+from ._merge_pr_cmd import merge_pr_app
+from ._prune_cmd import prune_app
 
 work_app = App(name="work", help="工作区管理")
