+++ packages/diyui.py/src/diyui/__init__.py
@@ -4,11 +4,11 @@ diyui - A thin reactive wrapper over UI frameworks.
 轻量响应式 UI 框架包装层。
 """
 
-from .base_app import BaseApp
-from .debug import DebugInfo, get_debug
-from .scheduler import ImmediateScheduler
-from .scope import ScopeConfig, ScopeNode
-from .signal import ScopeViolationError, Signal
+from ._base_app import BaseApp
+from ._debug import DebugInfo, get_debug
+from ._scheduler import ImmediateScheduler
+from ._scope import ScopeConfig, ScopeNode
+from ._signal import ScopeViolationError, Signal
 
 __version__ = "0.1.4"
 __all__ = [
