+++ packages/diyui.py/src/diyui/_scheduler.py
@@ -6,7 +6,7 @@ v0.1：ImmediateScheduler（同步执行/rerun 中延迟）。
 
 from collections.abc import Callable
 
-from . import signal as signal_mod
+from . import _signal as signal_mod
 
 
 class ImmediateScheduler:
