+++ packages/diyui.py/src/diyui/_scope.py
@@ -15,10 +15,9 @@ from dataclasses import dataclass
 from typing import TYPE_CHECKING, Any, TypeVar
 
 if TYPE_CHECKING:
-    from .base_app import BaseApp
+    from ._base_app import BaseApp
 
 C = TypeVar("C", bound="ScopeNode")
-
 
 @dataclass
 class ScopeConfig:
@@ -139,7 +138,7 @@ class ScopeNode:
         - staging 模式收集新 children，成功后原子替换，失败保留旧 UI
         - cell 执行期间 self 作为当前 context
         """
-        from . import signal as signal_mod
+        from . import _signal as signal_mod
 
         fn = self._cell_fn
         if fn is None:
@@ -163,7 +162,7 @@ class ScopeNode:
         signal_mod._current_cell_node = self
         signal_mod._rerun_depth += 1
 
-        from .debug import get_debug
+        from ._debug import get_debug
 
         debug = get_debug(self)
         debug.record_rerun()
