+++ packages/diyui.py/src/diyui/_signal.py
