+++ packages/diyui.py/src/diyui/providers/panel/pane/__init__.py
@@ -1,3 +1,3 @@
-from .markdown import Markdown
+from ._markdown import Markdown
 
 __all__ = ["Markdown"]
