+++ packages/diyui.py/src/diyui/providers/panel/widgets/__init__.py
@@ -1,5 +1,5 @@
-from .button import Button
-from .text_input import TextInput
-from .radio_button_group import RadioButtonGroup
+from ._button import Button
+from ._text_input import TextInput
+from ._radio_button_group import RadioButtonGroup
 
 __all__ = ["Button", "TextInput", "RadioButtonGroup"]
